coordinator: move BlockParameter to domain module
