+++ jvm-libs/core/domain-models/src/main/kotlin/net/consensys/linea/BlockParameter.kt
@@ -1,4 +1,4 @@
-package net.consensys.zkevm.coordinator.clients.smartcontract
+package net.consensys.linea
 
 sealed interface BlockParameter {
   fun getTag(): String
